refactor: update and include routes in endpoint directory for inPluginUpsells
